requirements.txt
Django>=1.6
psycopg2
Pillow
South
django-mptt